requirements_no_torch.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2024.6.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.4
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numba==0.61.2
numpy==2.1.2
onnxruntime==1.22.1
onnxruntime-gpu==1.22.0
openai==1.102.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.0.0
platformdirs==4.4.0
polars==1.32.3
pooch==1.8.2
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMatting==1.1.14
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rembg==2.0.56
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.16.1
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
tifffile==2025.8.28
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.187
ultralytics-thop==2.0.16
urllib3==2.5.0
